requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
pyngrok
google-generativeai